RRT/IRIS Tests + New Thing
Compared RRT and IRIS for both the maze and mattress-movers map. Also trying out this new implementation where in the case that Dijkstra fails (because IRIS fails to completely decompose the space, leading to regions of freespace), a forward+backwards Dijkstra can be run from start to goal and from goal to start and then batched RRT can be used to find the best (cheapest) path to connect both trees. Currently rewriting the inFreespace() and connectsTo() functions to be compatible with the VPolytope definition of the obstacles.