Two-way Dijkstra's more reliable [see description]
So the two-way Dijkstra for the mattress-movers problem is more reliable, in the sense that some of the batches stop freezing (AYYYY), but I noticed that it is lowkey pretty slow. The time sink comes from doing the tensor math/pytorch in the RRT function. I've been running this on CPU because the computer that I use to access Pydrake doesn't have an NVIDIA GPU, so I can't use cuda for Pytorch. I'm hoping that when this gets run on a GPU that it'll be faster. Also need to look into the point-sampling methods. Still not completely satisfied with how the RRT start-goal pairs are found.